fix(test): correct uint64 underflow in memory slope calculation
When HeapAlloc decreases between snapshots (normal GC behavior), the
uint64 subtraction underflows producing absurd slope values. Cast to
int64 before subtraction to handle negative deltas correctly.

Memory growth test results (50 req/s × 10 min, 30K requests):
- Heap growth: 0.39 MB (< 100 MB threshold)
- Memory trend: 0.10 MB between halves (plateau confirmed)
- No goroutine leaks, IP limiters cleaned up properly

+++ internal/server/memory_test.go
@@ -363,7 +363,7 @@ func runMemoryGrowthTest(t *testing.T, duration time.Duration) {
 		var lastNSlope float64
 		if lastN > 1 {
 			for i := len(snapshots) - lastN + 1; i < len(snapshots); i++ {
-				delta := float64(snapshots[i].HeapAlloc-snapshots[i-1].HeapAlloc) / (1024 * 1024)
+				delta := float64(int64(snapshots[i].HeapAlloc)-int64(snapshots[i-1].HeapAlloc)) / (1024 * 1024)
 				lastNSlope += delta
 			}
 			lastNSlope /= float64(lastN - 1)
